# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  student: 'student',
  company: 'company',
} as const;

export interface User {
  id: number;
  email: string;
  role: UserRole;
  name: string;
}

export interface AuthSession {
  user: User;
}

export type SignupInputRole = typeof SignupInputRole[keyof typeof SignupInputRole];


export const SignupInputRole = {
  student: 'student',
  company: 'company',
} as const;

export interface SignupInput {
  email: string;
  /** @minLength 6 */
  password: string;
  role: SignupInputRole;
  /** @minLength 1 */
  name: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface Job {
  id: number;
  companyId: number;
  companyName: string;
  companyInitials: string;
  title: string;
  description: string;
  requirements: string[];
  location: string;
  type: string;
  /** @nullable */
  salary: string | null;
  postedAt: string;
  applicantsCount: number;
  hasApplied: boolean;
}

export interface JobInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  requirements: string[];
  /** @minLength 1 */
  location: string;
  /** @minLength 1 */
  type: string;
  salary?: string;
}

export interface Profile {
  id: number;
  userId: number;
  name: string;
  headline: string;
  school: string;
  /** @nullable */
  graduationYear: number | null;
  skills: string[];
  bio: string;
  /** @nullable */
  resumeUrl: string | null;
}

export interface ProfileInput {
  /** @minLength 1 */
  name: string;
  headline: string;
  school: string;
  /** @nullable */
  graduationYear: number | null;
  skills: string[];
  bio: string;
}

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  pending: 'pending',
  reviewed: 'reviewed',
  interview: 'interview',
  rejected: 'rejected',
} as const;

export interface Application {
  id: number;
  jobId: number;
  jobTitle: string;
  companyName: string;
  studentId: number;
  studentName: string;
  studentEmail: string;
  /** @nullable */
  resumeUrl: string | null;
  status: ApplicationStatus;
  appliedAt: string;
}

export interface ApplicationInput {
  jobId: number;
  resume: Blob;
}

export type ApplicationStatusInputStatus = typeof ApplicationStatusInputStatus[keyof typeof ApplicationStatusInputStatus];


export const ApplicationStatusInputStatus = {
  pending: 'pending',
  reviewed: 'reviewed',
  interview: 'interview',
  rejected: 'rejected',
} as const;

export interface ApplicationStatusInput {
  status: ApplicationStatusInputStatus;
}

export type Applicant = Application;

export interface CompanyDashboard {
  jobs: Job[];
  totalApplicants: number;
  interviews: number;
  profileViews: number;
}

export type ListJobsParams = {
search?: string;
location?: string;
type?: string;
};