# lib/api-zod/src/generated/types/applicationStatusInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationStatusInputStatus } from './applicationStatusInputStatus';

export interface ApplicationStatusInput {
  status: ApplicationStatusInputStatus;
}