# lib/api-zod/src/generated/types/companyDashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Job } from './job';

export interface CompanyDashboard {
  jobs: Job[];
  totalApplicants: number;
  interviews: number;
  profileViews: number;
}