# lib/api-zod/src/generated/types/jobInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JobInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  requirements: string[];
  /** @minLength 1 */
  location: string;
  /** @minLength 1 */
  type: string;
  salary?: string;
}