# lib/api-zod/src/generated/types/signupInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SignupInputRole } from './signupInputRole';

export interface SignupInput {
  email: string;
  /** @minLength 6 */
  password: string;
  role: SignupInputRole;
  /** @minLength 1 */
  name: string;
}